Crate lux

Lux

Modules

color
Color creation functions and some named defaults.

font
Text drawing and loading functionality.

game
A game loop implementation using Lux for windowing and graphics.

graphics
All of 2d graphics related functionality.

interactive
Functionality for dealing with events, querying, and modifying the window.

modifiers
Many contexts can be modified and chained together in a logical way.

prelude
A collection of common traits, structs and functions that are recommended for average Lux usage.

types
Typedefs for use inside Lux.

window
The main window and the frames that are generated by the window.

Enums

LuxError
Any error that Lux might run into.

Type Definitions

LuxResult
A result returning either a value or a lux-generated error.