Module lux::color

Color creation functions and some named defaults.

Statics

ALICEBLUE
All of the wc3 named colors.

ANTIQUEWHITE

AQUA

AQUAMARINE

AZURE

BEIGE

BISQUE

BLACK

BLANCHEDALMOND

BLUE

BLUEVIOLET

BROWN

BURLYWOOD

CADETBLUE

CHARTREUSE

CHOCOLATE

CORAL

CORNFLOWERBLUE

CORNSILK

CRIMSON

CYAN

DARKBLUE

DARKCYAN

DARKGOLDENROD

DARKGRAY

DARKGREEN

DARKKHAKI

DARKMAGENTA

DARKOLIVEGREEN

DARKORANGE

DARKORCHID

DARKRED

DARKSALMON

DARKSEAGREEN

DARKSLATEBLUE

DARKSLATEGRAY

DARKTURQUOISE

DARKVIOLET

DEEPPINK

DEEPSKYBLUE

DIMGRAY

DODGERBLUE

FIREBRICK

FLORALWHITE

FORESTGREEN

FUCHSIA

GAINSBORO

GHOSTWHITE

GOLD

GOLDENROD

GRAY

GREEN

GREENYELLOW

HONEYDEW

HOTPINK

INDIANRED

INDIGO

IVORY

KHAKI

LAVENDER

LAVENDERBLUSH

LAWNGREEN

LEMONCHIFFON

LIGHTBLUE

LIGHTCORAL

LIGHTCYAN

LIGHTGOLDENRODYELLOW

LIGHTGREEN

LIGHTGREY

LIGHTPINK

LIGHTSALMON

LIGHTSEAGREEN

LIGHTSKYBLUE

LIGHTSLATEGRAY

LIGHTSTEELBLUE

LIGHTYELLOW

LIME

LIMEGREEN

LINEN

MAGENTA

MAROON

MEDIUMAQUAMARINE

MEDIUMBLUE

MEDIUMORCHID

MEDIUMPURPLE

MEDIUMSEAGREEN

MEDIUMSLATEBLUE

MEDIUMSPRINGGREEN

MEDIUMTURQUOISE

MEDIUMVIOLETRED

MIDNIGHTBLUE

MINTCREAM

MISTYROSE

MOCCASIN

NAVAJOWHITE

NAVY

OLDLACE

OLIVE

OLIVEDRAB

ORANGE

ORANGERED

ORCHID

PALEGOLDENROD

PALEGREEN

PALEVIOLETRED

PAPAYAWHIP

PEACHPUFF

PERU

PINK

PLUM

POWDERBLUE

PURPLE

RED

ROSYBROWN

ROYALBLUE

SADDLEBROWN

SALMON

SANDYBROWN

SEAGREEN

SEASHELL

SIENNA

SILVER

SKYBLUE

SLATEBLUE

SLATEGRAY

SNOW

SPRINGGREEN

STEELBLUE

TAN

TEAL

THISTLE

TOMATO

TURQUOISE

VIOLET

WHEAT

WHITE

WHITESMOKE

YELLOW

YELLOWGREEN

Traits

Color
A Color is any object that can be converted to a length-4 array of f32s.

Functions

hex_rgb
Convertes a u32 to a color by treating the last 3 bytes as rgb tripples.

hex_rgba
Convertes a u32 to a color by treating the 4 bytes as rgb tripples.

hsv
Constructs a color from Hue, Saturation and Value components.

hsva
Same as hsv but with an alpha component.

rgb
Constructs a color from R, G, and B components.

rgba
Constructs a color from R, G, B, and A components.