Enum lux::LuxError

pub enum LuxError {
    WindowError(String),
    OpenGlError(String),
    ImageError(ImageError),
    ShaderError(ShaderError),
    IoError(IoError),
    FontNotLoaded(String),
    DrawError(DrawError),
    TextureCreationError(TextureCreationError),
    IndexBufferCreationError,
    VertexBufferCreationError,
}

Any error that Lux might run into.

Variants

WindowError

An error that can occur when creating a window.

OpenGlError

An error that can occur when creating an opengl context.

ImageError

An error related to image decoding.

ShaderError

An error that can occur when compiling or linking shaders.

IoError

An error that can occur when required I/O fails.

FontNotLoaded

An error that can occur when attempting to use a font that hasn't been loaded yet.

DrawError

An error that was produced while submitting a draw call

TextureCreationError

An error occured while creating a texture

IndexBufferCreationError

An error creating an index buffer occured

VertexBufferCreationError

An error creating an vertex buffer occured