Struct lux::game::Loader

pub struct Loader<G> {
    // some fields omitted
}

A struct that stores closures which load assets and prepare the game for running.

Methods

impl<G> Loader<G>

fn do_async<R, FL, FA, S: Into<String>>(&mut self, desc: S, load: FL, apply: FA) where FL: Fn() -> LuxResult<R> + Send + 'static, FA: Fn(R, &mut Window, &mut G) -> LuxResult<()> + 'static

Performs part of a job asynchronously and then applies the result of that loading to the game.

The two closures are separate parts of the same loading job.

load: Does the bulk of the loading work. Returns a LuxResult with a return value.

apply: Takes the return value and applies it to the Window and Game object.

The load closures are all run in separate threads, but the apply closures must run sequentially in the same thread in order to share the Game and Window objects.

fn do_sync<F, S: Into<String>>(&mut self, desc: S, load: F) where F: Fn(&mut Window, &mut G) -> LuxResult<()> + 'static

Performs a job without an asynchronous component.

The closure is executued on the main thread, and is called along with the apply closures from the do_async method.