Module lux::graphics

All of 2d graphics related functionality.

Most of the methods that actually let you draw things are defined in two main traits: Canvas and PrimitiveCanvas.

The two objects that can be drawn to are are Frame and DrawableTexture.

Reexports

pub use glium::index::PrimitiveType::*;

Structs

ColorVertex
A colored vertex.

ContainedSprite
A sprite that can be drawn to the screen.

DrawableTexture
A texture that can be drawn to like a regular canvas.

Ellipse
An ellipse that can be drawn to the screen.

NonUniformSpriteSheet
A uniform sprite sheet is a sprite sheet that has parts of the original sprite broken up into chunks that are associated to a Key.

Rectangle
A Rectangle that can be drawn to the screen.

Sprite
A ref-counted reference to a texture on the GPU.

TexVertex
A textured vertex.

Texture
An owned texture on the hardware.

UniformSpriteSheet
A uniform sprite sheet is a sprite sheet that is broken up into a grid of equally sized sub-sprites.

Enums

PrimitiveType
List of available primitives.

StencilType
When modifying a stencil or clearing the stencil buffer, operations require a StencilType.

Traits

Canvas
Canvas is the main trait for drawing in Lux. It supports all operations that paint to the screen or to a buffer.

IntoSprite
Implemented by any object that can be converted into a Sprite.

PrimitiveCanvas
A Primitive canvas is a trait that is implemented by objects that can have draw commands issued to them.

TextureLoader
TextureLoader is implemented on any object that can load textures.

Implementations