Struct lux::graphics::TexVertex

pub struct TexVertex {
    pub pos: [f32; 2],
    pub tex_coords: [f32; 2],
}

A textured vertex.

tex_coords is the position on the texture where x and y are in the range 0.0 to 1.0.

Fields

pos

The position in screen space

tex_coords

The texture cooordinates [x, y] where x and y are in the range 0.0 to 1.0.