Module lux::interactive

Functionality for dealing with events, querying, and modifying the window.

Modules

keycodes
A keycode is a platform independent way to refer to keys on the keyboard.

Structs

EventIterator
An iterator for windowing events.

Enums

Event
An even coming from an Interactive object.

MouseButton
A handy enumeration for the buttons on a mouse.

Traits

AbstractKey
A conversion trait for representing the different ways that a key can be represented.

Interactive
A trait for objects that are interactive to the user. The only known impelementation for this trait is the glutin Window.