Struct lux::window::Frame

pub struct Frame {
    // some fields omitted
}

A frame is a render target that can be drawn on.

Because frame rendering will wait on vsync, you should - as the name implies - use one Frame instance per frame.