Struct lux::window::Window

pub struct Window {
    // some fields omitted
}

A 1 to 1 correlation with a window shown on your desktop.

Lux uses Glutin for the window implementation.

Methods

impl Window

fn assert_no_error(&self)

Panics if an OpenGL error has occurred.

fn new() -> LuxResult<Window>

Constructs a new window with the default settings.

fn restock_events<I: DoubleEndedIterator<Item=Event>>(&mut self, i: I)

Add the events from an iterator of events back to the internal event queue.

fn process_events(&mut self)

Query the underlying window system for events and add them to the the interal event queue.

This function is automatically called by is_open(), events().

fn cleared_frame<C: Color>(&mut self, clear_color: C) -> Frame

Produce a frame that has been cleared with a color.

fn frame(&mut self) -> Frame

Produce a frame that has not been cleared.