Struct lux::window::WindowOptions

pub struct WindowOptions {
    pub dimensions: (u32, u32),
    pub title: String,
    pub vsync: bool,
    pub multisampling: u16,
    pub transparent: bool,
    pub decorations: bool,
}

A set of options that can be applied to a window

Fields

dimensions

The size of the window in pixels.

title

The title displayed on the top of the window.

vsync

If Vsync is enabled

multisampling

The number of multisampling passes. Must be a power of 2.

transparent

True if the window should be transparent.

decorations

True if the window should have no border or title-bar.