Struct lux::game::GameRunner

pub struct GameRunner<G: Game> {
    pub window: Window,
    pub game: G,
    // some fields omitted
}

A struct that wraps a Game and a Window and implementes a game loop.

Fields

window

The window being used to run the game.

game

The wrapped game.

Methods

impl<G: Game> GameRunner<G>

fn new(game: G) -> LuxResult<GameRunner<G>>

Constructs a new GameRunner wrapping a game.

Attempts to build a window as well.

fn step(&mut self) -> LuxResult<()>

Moves the game forward one "step".

A "step" has two main phases

Update Phase: The Games update() function is called any number of times (including 0).

Render Phase: The Games render() function is called exactly once.

The game runner keeps track of lag manually, so update() might be called more than once per step. If you are calling step manually and you stop the game (for example, to pause the game), you must call reset_lag so that step doesn't think that the game lagged for a very long time.

fn run_until_end(&mut self) -> LuxResult<()>

Runs the game until the game is terminated.

The game loop is fairly complicated but is explained as the final example in this blog post: http://gameprogrammingpatterns.com/game-loop.html

Main features

All updates happen in discrete time steps.

Multiple updates can happen sequentially in order to handle lag and get back on step.

Multiple renders can happen sequentially if the computer is too fast. Renders will keep occuring until enough time has built up to trigger another update.

The amount of "lag" time is passed in to the render function so that the user can accomodate for lag.

fn reset_lag(&mut self)

Resets the stored lag.

This function should only be called when resuming calls to step after not calling it regularly.

fn do_load(&mut self) -> LuxResult<()>

Execute the game loading closures that would be generated by Game::load.